+++ sdgym/__init__.py
@@ -8,10 +8,11 @@ __author__ = 'MIT Data To AI Lab'
 __copyright__ = 'Copyright (c) 2018, MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
 __license__ = 'MIT'
-__version__ = '0.3.1.dev2'
+__version__ = '0.3.2.dev0'
 
 from sdgym import benchmark, results, synthesizers
 from sdgym.benchmark import run
+from sdgym.collect import collect_results
 from sdgym.datasets import load_dataset
 
 __all__ = [
@@ -20,4 +21,5 @@ __all__ = [
     'results',
     'run',
     'load_dataset',
+    'collect_results'
 ]